Introduce Link class in uml package
